initial commit
Inspired by the Mission Impossible series and leveraging my interest in Python programming, I created a project that assigns tasks to Ethan Hunt and other members of the IMF team according to the movie's scenarios. This project showcases my ability to integrate pop culture with coding to make learning and development more engaging.